+++ packages/nimble-components/src/table/template.ts
@@ -9,7 +9,7 @@ import { TableRow } from './components/row';
 export const template = html<Table>`
     <template role="table">
         <div class="table-container">
-            <div role="rowgroup">
+            <div role="rowgroup" class="header-container">
                 <div class="header-row" role="row">
                     ${repeat(x => x.columnHeaders, html<string>`
                         <${DesignSystem.tagFor(TableHeader)} class="header">